internal/syscall/unix: add PidFDSendSignal for linux
CL 520266 added pidfd_send_signal linux syscall numbers to the
syscall package for the sake of a unit test.

As we may need to use pidfd_send_signal from e.g. os package, let's
revert the changes to syscall package, add the pidfd_send_signal
syscall numbers and the implementation to internal/syscall/unix,
and change the above test to use it.

For #51246

Change-Id: I862174c3c1a64baf1080792bdb3a1c1d1b417bb4