os: add pidfd support for linux
DO NOT REVIEW
DO NOT MERGE
(this is a squashed commit for a WIP branch; for testing only)

> internal/syscall/unix: add PidFDSendSignal for Linux

Currently only used by a test in syscall package.

> syscall: make StartProcess return pidfd on Linux

Reuse the handle variable for it.

Also, ensure that pidfd is closed in Release.

> os: use PidfdSendSignal on linux if possible

System call pidfd_send_signal is available since Linux 5.1, and the
ability to get pidfd from clone syscall is there since Linux 5.2.
So, if pidfd is set, we can be sure pidfd_send_signal is available
(unless disabled by e.g. seccomp security policy).

Use it, if available, and if it works, set the pidfdSendSignalWorks
which tells we no longer have to care about the wait/kill race.

> os: amend FindProcess to use pidfdOpen on linux

Change-Id: I40ea5742a4dca896c13dc5310f4517ba671799aa