runtime/internal/atomic: add 386/amd64 And/Or operators
This CL adds the atomic primitives for the And/Or operators on x86-64.

Note that the race variants for those operators are not yet implemented
since we depend on an upstream llvm tsan patch as well as rebuilding the
race runtime with x/build/cmd/racebuild. This will come as a separate patch
at a later time once the infraestructure and upstream patches supporting it
are ready.

See llvm/llvm-project#65695 for the llvm tsan patch.

For #61395

Change-Id: I23ef5192866d21fc3a479d0159edeafc3aeb5c47
GitHub-Last-Rev: 55143c468bf1c1294ad7625d28c5735959fa88e3
GitHub-Pull-Request: golang/go#62621