reflect: build interfaces without leak to heap in Value.Interface
WIP. Feedback welcome.

This is part of a series of CLs that aim to reduce how often interface
arguments escape for the print functions in fmt.

Those functions currently make use of Value.Interface, which leaks its
Value.ptr to the heap. Previously, it would leak to the heap for two
reasons. Our immediately prior CL attempts to address one reason (in
packEface), and this CL attempts to sidestep the second reason.

Value.Interface is a thin wrapper around valueInterface, which
calls makeMethodValue if handed a method value like
ValueOf(Foo{}.Method). In makeMethodValue, a *methodValue is
always allocated, and it stores our Value.ptr:

  &methodValue{...} escapes to heap in makeMethodValue:
    flow: fv ← &<storage for &methodValue{...}>:
      from &methodValue{...} (spill) at .\makefunc.go:113:8
      from fv := &methodValue{...} (assign) at .\makefunc.go:113:5
    flow: ~r0 ← fv:
      from Value{...} (struct literal element) at .\makefunc.go:129:14
      from return Value{...} (return) at .\makefunc.go:129:2

The data-flow graph built by escape analysis is a simplified model
that ignores conditionals, so even though makeMethodValue is only
conditionally called by interfaceValue and hence the *methodValue
is only conditionally allocated at run time, escape analysis concludes
at compile time that the Value.ptr handed to Value.Interface
always escapes:

  parameter v leaks to <heap> in valueInterface with derefs=0:
    flow: <heap> ← v:
      from makeMethodValue("Interface", v) (call parameter) at .\value.go:1511:22

To avoid this, we cheat a bit by defining a new function with equivalent
functionality to makeMethodValue except that it panics if handed a
method value and never calls makeMethodValue, thus avoiding the
Value.ptr escaping to the heap. The Value.ptr now instead leaks to the
result interface, which is what we want.

We add an allocation test for our new function, but no one otherwise
calls our function in this CL. A future CL will make use of it from the
fmt package.

Change-Id: I7d635eb226aff45b9d752616d770222071700f1d