runtime/internal/atomic: add 386/amd64 And/Or operators
This CL adds the atomic primitives for the And/Or operators on x86-64.
It also includes missing benchmarks for the ops.

Note that the race variants for those operators are not yet implemented
since we depend on an upstream llvm tsan patch as well as rebuilding the
race runtime with x/build/cmd/racebuild. See https://github.com/golang/go/issues/62624.

The patch to update the race detector will come at a later time once the
infraestructure and upstream patches supporting it are ready.

For #61395

Change-Id: I23ef5192866d21fc3a479d0159edeafc3aeb5c47
GitHub-Last-Rev: 1f1a479b9763586e8ac0ea32e06a7a6da8657e09
GitHub-Pull-Request: golang/go#62621