cmd/compile/internal/escape: support a new //go:leaktoresult pragma
WIP. Feedback welcome.

TODO: is this a bad idea?
TODO: file an issue for discussion to see if desirable.

This is part of a series of CLs that aim to reduce how often
interface arguments escape for the print functions in fmt.

Currently, using //go:linkname defaults to causing function
arguments to escape. This can be overridden with //go:noescape,
but in some cases, the correct answer is that a function
leaks input parameters to the function results.

Provide a pragma that allows supplying that answer. We limit
it to use within the standard library.

The pragma currently records that all input arguments leak to
all output results, ignoring any scalar inputs or results.
In theory, it could be extended for more fine-grained control,
but we do not currently need any additional control, including
in our journey to help the print functions in fmt, our
most immediate target has a single input and a single output.

We also update some of our basic escape analysis tests,
though it is only our next CL that exercises this pragma,
including with an allocation test.

Change-Id: I093569bad555ecc4fb0e3da76a2d4880f09afbac