math: added generic and variadic MaxMany and MinMany functions
Although the Min() and Max() functions in the math package are usable, they have 2 main drawbacks. First, the arguments must be float64 types, which makes the user cast their other numeric types to float64 like int before using the function. Second, the number of arguments is limited by two. If an user wants to find the maximum or minimum value among multiple arguments, he/she would have to implement in-house solutions.

A possible solution is to add a couple functions like these MaxMany() and MinMany(), which are generic and variadic -- you may pass in any number data type argument, and as many as you like. An interface named Number with type constraints to all ints, uints, and floats, was added to help making the function generic.

Max() and Min() cover specific cases like dealing with infinities and NaN, which were also implemented in MaxMany() and MinMany(). The Signbit() case is not covered because of unfamiliarity with the topic, but can be added if required.

The functions were thoroughly tested and the test cases were added to the all_test.go file.

Change-Id: I785725dac579436f3180ed99d789df50c4a31e1c
GitHub-Last-Rev: 2b495ad0ab431f2f755ea3233209842cc70ae3b5
GitHub-Pull-Request: golang/go#62696