cmd/compile: avoid ANDCCconst on PPC64 if condition not needed
In the PPC64 ISA, the instruction "and immediate" is only
available as one which sets CR0. This means CR0 is being set
unnecessarily in some cases. Some compiler passes check for flags
and generate different code based on that.

In those cases when the constant used by the ANDCC is a right
justified consecutive set of bits, a equivalent shift
instruction can be used which doesn't clobber CR0.

This rule has been added to the late rules file.

Some codegen tests had to be updated.

Change-Id: I304f607c039a458e2d67d25351dd00aea72ba542