cmd/compile: avoid ANDCCconst on PPC64 if condition not needed
In the PPC64 ISA, the instruction an "and immediate" is only
available as one which sets CR0. This means CR0 is being set
unnecessarily, and affects other passes which check for that.

In those cases when the constant used by the ANDCC is a right
justified consecutive set of bits, another shift instruction can
be used. This rule has been added to the late rules file.

Some codegen tests had to be updated.

Change-Id: I304f607c039a458e2d67d25351dd00aea72ba542