runtime: check for illegal heap to stack pointers during GC
WIP. Feedback welcome. (Currently fails in reflect.TestGCBits, which
seems to be a special test that reaches into the runtime).

This is an attempt to take advantage of Cherry Mui's excellent unmerged
CL 408827.

This CL updates that CL for some changes since then (e.g., due to
goexperiment.AllocHeaders), as well as adds checking debug.invalidptr,
adjusts the offset printed, and dumps more information when an illegal
pointer is detected.

If we purposefully introduce a subtle escape analysis bug, we can see
this trigger.

TODO: it would be nice to commit a test that triggers this. The
top-level test directory harness does not seem to have support for a
flag that means "expect failure", but perhaps we could add that, or
perhaps there is a more natural place to test this.

Change-Id: Ife002888a2fe3a5d3393b60332c89e43a05026b7