cmd/compile/internal/inline: support reading adjustments from config file
Change the inline heuristic package to score scoring adjustments in
tabular form, and add debug options to write out JSON for the table or
read in the contents of the table from a JSON file. This is intended
to support doing performance experiments that change the score values
for specific heuristics.

Change-Id: Ibd1ee334684af5992466556a69baa6dfefb246b3