strings: added function to get string pointer from a string constant
This function will be useful mostly during testing. When we want to pass a string pointer to a struct member each time we have to create a separate variable and pass the address of that to the member variable, Which is a pain. This wrapper function will ease all those efforts.
Eg:
type Employee struct{
Name *string
}

To pass a variable in testing , we have to
testName  := "hello"

emp := Employee{
  Name: &testName,
}

With this function it is

emp: = Employee{
  Name: strings.StringPointerFrom("hello")
}

It is really helpful when we are using table tests.

Change-Id: I93f719200a48c509c90eb225a7809f5b26a02d08
GitHub-Last-Rev: 5fd0499b9ef636c34421f6ab2cce6c581776d167
GitHub-Pull-Request: golang/go#63299