cmd/go: add support for mod tools
Running `go tool` with no arguments will now list built in tools
followed by module defined tools.

Running `go tool X` where X matches either the full package path,
or the last segment of the package path, of a defined tool will
build the tool to a known location and immediately execute it.

This change adds a new cached directory which for now is excluded from
GOEXPERIMENT=cacheprog in the same manner as the fuzz cache.

Change-Id: I02249df8dad12fb74aa244002f82a81af20e732f